Remove factory + moved js to assets
