+++ plugins/MauticFocusBundle/EventListener/CampaignSubscriber.php
@@ -20,7 +20,6 @@ use Mautic\CoreBundle\EventListener\CommonSubscriber;
 use Mautic\PageBundle\Helper\TrackingHelper;
 use MauticPlugin\MauticFocusBundle\FocusEvents;
 use MauticPlugin\MauticFocusBundle\Model\FocusModel;
-use Symfony\Bundle\FrameworkBundle\Routing\Router;
 use Symfony\Component\Routing\RouterInterface;
 
 class CampaignSubscriber extends CommonSubscriber
@@ -82,6 +81,7 @@ class CampaignSubscriber extends CommonSubscriber
             'description'            => 'mautic.focus.campaign.event.show_focus_descr',
             'eventName'              => FocusEvents::ON_CAMPAIGN_TRIGGER_ACTION,
             'formType'               => 'focusshow_list',
+            'formTheme'              => 'MauticFocusBundle:FormTheme\FocusShowList',
             'formTypeOptions'        => ['update_select' => 'campaignevent_properties_focus'],
             'connectionRestrictions' => [
                 'anchor' => [
